Add new GitGud.Authorization module
In order to have a more generic way of handling resource authorizations,
the new module provides authorized?/3 and enforce_policy/3.

Both functions make use of the new GitGud.AuthorizationPolicies protocol
to find out if a user can perform an action on a given resource.

GitGud.Repo does not have can_read/2 and can_write/2 but implements the
new protocol for :read and :write actions.