Add helper navigation_item/4 to layout view
A navigation_item renders a <li></li> block and sets the tag class to
"is-active" if the given route helper and action matches.

This is very nicely implemented =)

Based on GitGud.Web.Router.__helpers__/0 we generate a function for each
available :helper. This way we have no runtime costs for resolving
Phoenix controllers from helpers.

The current_route?/3 function also provides :only and :except options:

* current_route?(conn, :repository, only: [:show, :edit])
* current_route?(conn, :repository, except: [:blob, :tree])