Refactor Git receive-pack service
Instead of letting the GitRekt.WireProtocol.ReceivePack module apply the
packfile, we now have a :callback where we can delegate these actions to.

Here's example to let GitGud.Repo apply the packfile to the repo:

    Service.new(exec, callback: {Repo, :push, [repo, user]})

The GitGud.Repo push/3  function is responsible for applying packfile
objects to the repository. It must also apply reference changes.

See #10 for more details.