Add support for cursor based pagination
The new GitGud.Web.View.paginate_cursor/5 function should help with
issue #36 by using a *cursor* (a commit oid for example) instead of
having to process the entire list of items.

Here's a small example on how it's used for commit history:

    page = paginate_cursor(@conn, @commits, &(oid_fmt(&1) == &2), &oid_fmt(&1.oid))

The third argument is used to filter matching items. The fourth argument
is used to return a *cursor* string for the next or previous page.

We can see huge performance increase (up to 1000x faster) when fetching
the history for a given tree or blob. In some cases, it still does not
perform much faster than paginate/3. This is due to the fact that if a
file has not changed muched, we have to process the stream as long as we
try to aggregate `limit`+1 items.