Add support for password reset
Basically, when a user forgets his password, he/she can request a
request password token sent by email. The token ensures that he/she is
the owner of the account. Once clicked, the link will reset the password
(set it to NULL in the database) and allows the user to set a new one in
his user settings.