Add views helpers to format Git log messages
The commit_message_format/2 function splits the title from the body of
the message and formats the body to ensure that each line is wrapped
correctly. You can pass the :wrap option to force to wrap after a number
of characters. If :wrap is ommited, the function will wrap using the
default message formatting.