Add GitRekt.GitAgent
GitAgent provides a high-level API for calling functions on Git
repository. It is meant to replace the Git helper modules and functions
available in the :gitgud app. it provides a simple wrapper around
GitRekt.Git functions and allows to call them from the current process
or a *proxy* GenServer process.

Basically, to use the module synchronously, just call:

    {:ok, pid} = GitAgent.open("path/to/repo.git")

In contrast, to wrap it in a dedicated GenServer, call:

    {:ok, ref} = GitAgent.start_link("path/to/repo.git")

Once you get a pid or a reference from one of the above functions, you
can execute commands using the same API.