Add new schema GitGud.GitCommit
The new schema is used to store meta informations about Git commits.
Currently, only the oid and parents are stored but this might expand to
other commit header properties in future.

Each time a repository is pushed (see Repo.push/2) the receive-pack is
parsed and each commit object is written into the GitCommit table.