Add batching resolution for Git delta objects
Using the libgit2 workflow for storing a Git pack object is not
perfoming well with lots of objects.
For each :delta_reference object, libgit2 has to lookup the ODB for the
base oid first, build the new object and write it back to the ODB.

The new GitRekt.ReceivePack.resolve_pack/1 function resolves the
delta-reference objects based on the avaible data in the pack.
In other words, it tries to resolve as much delta-references as possible
without having to lookup the ODB.

Using the :postgresql backend, we then have the possibility to batch
insert all the resolved objects at once.