Use paginate_cursor/5 with commit slices
Instead of running two queries (one for fetching the history page and one
for the total count of commits), we use paginate_cursor/5.

Passing a list to paginate_cursor/5 will use a different implementation
than when passing a stream.

If the list contains limit+1> items, the "Next" button will be enabled.
The "Previous" button is enabled if conn.path_params["p"] > 1.