Add GPG key verification when listing commits
This is a first attempt to retrieve and show GPG signed commits when
listing commits. GitGud.Web.CodebaseView.batch_commits/2 makes an
additional query to fetch users matching the commit's GPG keys-id.

The commit is only signed if the GPG key matches the author. If a GPG
key is matched but is not associated to the commit's author, we don't
show the *verified* icon.