Fetch last commits when browsing repositories
GitRekt.GitAgent.tree_entry_with_commit/3 returns the last commit of for
a given path.

GitRekt.GitAgent.tree_entries_with_commits/3, returns a list of tree
entries and their last commit for a given path.

Note that finding the last commit for a given pathspec is significantly
slow. We will try to optimize things in future.

See #52 for more details.