Stream request body in Git smart http backend
Instead of reading the entire body at once and use WireProtocol.run/2 to
execute the git-*-pack command, we stream the Plug.Conn body and use
WireProtocol.exec/2 in the same way we do in SSHServer.

In the previous implementation, we did :gzip.unzip/1 the full body if
gzip was set as 'content-encoding' http header. Not sure how the new
implementation behave with compressed requests data. It works with the
lastest Git version on MacOS. I hope :cowboy handles this automatically.

Closes #55.