Fix warning in GitGud.Web.OAuth2View
Phoenix resolve the template path for OAuth2View to o_auth2, we used to
call Phoenix.Template.__using__/1 to set the :path option. Because the
macro is already called when using GitGud.Web, :view, we got a warning.

It's now possible to pass a {atom, keyword} tuple to __using__/1
allowing to forward options to the underlying module.