Add GitGud.RepoSupervisor
The repository supervisor combines a Registry and a DynamicSupervisor.

Both are used to store and retrieve GitRekt.GitAgent processes. In order
to start a new agent use start_agent/1. Too retrieve a repo use lookup/1.

The GitRekt.GitRepo implementation for GitGud.Repo uses start_agent/1 in
its put_agent/1 callback. If a process with the same name is already
started, the PID of the process is returned.