Add GitGud.RepoPoolMonitor
The pool monitor is used to monitor GitRekt.GitAgent processes.

It retains shared ownership to Git agent processes. Several client
processes may own the same agent. Once all client processes are
terminated, a timer kills the agent after 30_000ms. If a new clients
claims ownership in this time period, the timer is canceled and the
agent is kept alive.