Remove GitGud.Repo.init_agent/2
Instead we always use GitGud.RepoPool in order to retrieve the
repository Git agent. In combination with previous commit, this allows
to have a dynamic supervisor for spawning and retrieving repositories
that stops automatically after a certain @idle_timeout.

Not sure if we need GitGud.GraphQL.RepoMiddleware anymore. Changing
affected resolvers to provide :repo as parent should do the trick.