Add JS functions to retrieve authenticated user
We now have the const currentUser (in auth.js) that returns the login
and avatarUrl of the authenticated used. This searches the DOM tree for
for #viewer-login and #viewer-avatar elements and parse them.

We also have a new GraphQL root query called viewer() returning the
authenticated user executing the query or null if not authenticated.