Change authorization for empty repositories
Instead of listening and rendering empty repositories to everyone, we
now check if :pushed_at != nil additionally to checking :public == true.

This means that a public repository that has not been pushed once will
not be listened or renderer by default. Maintainers are still able to
see such a repository.