Refactor Telemetry events for caching
Instead of passing the cache args, we pass the Git operation with args.
This way we can print the same op-args combo in the same way we do it
for [:gitrekt, :git_agent, :call] Telemetry events.

We also do not use the [:gitrekt, :git_agent, :cache] anymore but three
different event names instead.