Better handling of streams in GitRekt.GitAgent
Caching streams now really depends on the cache adapter implementation.
For example the result of GitAgent.tree_entries/2 and commit_parents/2
varies depending on the cache implementation:

- if make_cache_key/1 returns nil for the operation, it returns a stream.
- if it returns a cache key, it returns Enum.to_list(stream).

references/3 and history/3 are to special cases which are handled
differently. references/3 always returns a stream while history/3
tries to cache the result if :stream_chunk_size is set to :infinity.

Also a default :stream_chunk_size is now configurable via child specs.