Add repositories contributors
The repositories_contributors table keeps tracks of users who have
contributed to a repositories. Currently the only way to contribute to a
repository is to contribute code to the Git repository.

GitGud.RepoStorage.push/3 and push_commit/4 will search for registered
users based on the commits author and committer emails and add them to
the repository contributors.