Refactor caching in GitRekt.GitAgent
Additionally to the internal caching mechanism, we now have the
possibility to create the cache outside the Git agent. This provides
support for caching when running Git commands on a GitRekt.Git.repo
handle.

The new external caching mechanism also provides benefits when running
against a PID agent. By having access to the cache outside of the agent,
we don't have to rely on message passing for retrieving results from the
cache.

The GitRekt.GitRepoprotocol implementation for GitGud.Repo now supports
the new mechanism in order to keep track of the ETS cache table along
the agent.