Detach cache behaviour from Git agent GenServer
Instead of having the (ETS) cache as part of the GenServer state, we
don't rely on the cache to be known internally. Initializing the cache
is done in the GitGud.Repo implementation of the GitRekt.GitRepo
protocol. By using :ets.give_away/3 we allow the agent to write while
other processes only have read access.

Telemetry events diverge a tiny bit when using caching, if the operation
hits the cache successfully, cache: true will be passed to the handler.