Refactor GitGud.Web.CommitDiff
Kind of messy but that's what we have so far:

* CommitDiffLive - LiveView rendering a commit diff.
* CommitLineReviewLive - LiveComponent rendering a line review.
* CommitDiffDynamicFormsLive - LiveComponent rendering dynamic forms.
* CommitDiffDynamicReviewsLive - LiveComponent rendering dynamic reviews.

Basically CommitDiffLive is rendered only once. By using components for
all moving parts we can skip re-renders and using send_update/2 for
interacting with reviews, forms and comments.

Both CommitDiffDynamicFormsLive and CommitDiffDynamicReviewsLive have JS
hooks in order to move new components to the right table row.