Refactor all Ecto schemas
Field that are not meant to be set by the user are not taken into
account in the associated changeset/2. Those fields are given directly
when inserting into the database.

For example GitGud.Issue.create/4 function is meant to be used as follow:

  GitGud.Issue.create(
    repo,
    author,
    title: "Help",
    comment: %{body: "I need help!"}
  )

This helps with security concerns, and makes the changesets less
cumbersome to use and reduce complexity when validating forms with
Phoenix.LiveView.

Some minor changes have been done to the DB migrations...