Implement git agent pool using :nimble_pool
This is very experimental. The supervisor initializes a bunch of
C NIF handles reprensenting a single repository. The supervisor also
stores a public cache for the repository.

By calling GitGud.RepoPool.checkout/3, the current process acquires one
of the available repository handles.

It allows greater performanes for one-off operations but is not adequate
for long running processes such as LiveViews.