Refactor GitRekt.WireProtocol.new/3
Instead of taking a struct implementing the GitRekt.GitRepo protocol,
the first argument has to be a GitRekt.GitAgent.agent type.

Both transport protocols (HTTP, SSH) use GitRekt.GitAgent.start_link/1
for this purpose. This means that a dedicated agent is started for each
incoming git-upload-pack and git-receive-pack command.

We do not rely on GitGud.RepoPool here. See #68 for more details.