Refactor git-receive-pack implementation
We now have a big, fat and clumsy receive-pack implementation 😅

When receiving a PACK file, the implementation does produce a few
different things:

1) A list of objects and delta objects parsed with GitRekt.Packfile.
2) A binary representing the raw PACK file.
3) A writepack ready to be committed.

GitRekt.WireProtocol.ReceivePack.apply_pack/2 provides different ways
for writting the incoming PACK to the ODB:

1) :write -- Write each (loose) object available in a loop.
2) :write_pack -- Write the raw PACK file to the disc and generate idx.
3) :commit_pack -- Commits the already written PACK file.

So currently we have three different ways for writing incoming data from
a git-receive-pack command.

I think we should stick to 3) and get rid of 1) and 2). This way we
would not need to keep the entire PACK in memory as it would be streamed
directlty to disk.

See #68 for more details.