Refactor git-receive-pack implementation
We only use the writepack and do not parse the Packfile ourselves.
Doing this makes the whole process much faster as we do write the
incoming Git PACK file directly to disk and let libgit2 index it.

Doing so, we loose the ability to parse the objects contained in the
PACK and cannot do further actions (aggregate contributors, reference
issues from commit messages, etc.). We will add these features later on.