Add JS hook global-search.js
Because GitGud.Web.GlobalSearchLive is rendered in almost every page,
we have a special mechanism for only connecting the LiveSocket when
appropriate.

Normally, the live-views.js hook connects the LiveSocket whenever a
Phoenix LiveView is present in the DOM. Now if is not the case anymore
when data-phx-connect-later is set. In this case, the LiveView must
connect the socket manually.