Keep track of NIF references in GitRekt.GitAgent
When calling GitRekt.GitAgent from other nodes, NIF resources might be
garbage collected between two calls:

    {:ok, head} = GitRekt.GitAgent.head(agent)
    {:ok, commit} = GitRekt.GitAgent.peel(agent, head) #1
    {:ok, message} = GitRekt.GitAgent.commit_message(agent, message) #2

In the above example, between #1 and #2 the agent process might garbage
collect the commit. This could lead lead to a bad argument error in #2.

In order to counter this, GitRekt.GiAgent stores a set of borrowed NIF
resources.  This way the resources are reference counted and available
between consecutive calls. Note that the references are dropped once a
:DOWN message is received from a given client process.