Add support for distributed setup with :global
The new GitGud.Repo :volume field introduced in the last commit plays a
major part for finding out on which node a repository stores it's data.

A storage server is started on each node at startup. In the init/1
callback, it will make sure that the :git_root folder is tagged with a
volume ID (random 32 bit string). This ID is used across the cluster to
retrieve a set of repositories. Note that the repository :volume field
is assigned to local volume per default.

* Add :libcluster and start Cluster.Supervisor on application startup.

* GitGud.RepoPool is a GenServer embedding a DynamicSupervisor, a local
  Registry and a global ETS table for round-robin allocation of agents.

* GitGud.RepoStorage is a GenServer able to work on the right volume in
  a distributed setup. A storage server is started as part of the
  application supervision tree and ensure that file operations are executed
  on the right node.

* Remove GitGud.RepoMonitor. We might implement a different behaviour
  for pool live-time management and allocation later on.